refactor: move typography into utils
