+++ .storybook/addons.js
@@ -1,2 +1,3 @@
-import '@storybook/addon-actions/register';
-import '@storybook/addon-links/register';
+// import '@storybook/addon-actions/register'
+import '@storybook/addon-links/register'
+import '@storybook/addon-knobs/register'
