+++ .storybook/config.js
@@ -1,6 +1,6 @@
 import { configure, addDecorator } from '@storybook/react'
 import StoryWrapper from './StoryWrapper'
-import typography from '../src/typography'
+import { typography } from '../src/utils'
 
 // automatically import all files in src ending in *.stories.js
 const req = require.context('../src', true, /.stories.js$/)
@@ -15,4 +15,6 @@ typography.injectStyles()
 
 // wrap all stories with StoryWrapper
 addDecorator(StoryWrapper)
+
+// load stories
 configure(loadStories, module)
