+++ package.json
@@ -23,13 +23,14 @@
     "typography": "^0.16.17"
   },
   "devDependencies": {
-    "prettier": "^1.12.0",
-    "@storybook/react": "^4.0.8",
+    "@babel/core": "^7.1.6",
     "@storybook/addon-actions": "^4.0.8",
+    "@storybook/addon-knobs": "^4.0.8",
     "@storybook/addon-links": "^4.0.8",
     "@storybook/addons": "^4.0.8",
-    "@babel/core": "^7.1.6",
-    "babel-loader": "^8.0.4"
+    "@storybook/react": "^4.0.8",
+    "babel-loader": "^8.0.4",
+    "prettier": "^1.12.0"
   },
   "keywords": [
     "gatsby",
