+++ src/components/LinkListItem/index.stories.js
@@ -1,8 +1,13 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
-import { action } from '@storybook/addon-actions'
+import { withKnobs, text, boolean } from '@storybook/addon-knobs'
 import LinkListItem from './index'
 
-storiesOf('LinkListItem', module).add('default', () => (
-  <LinkListItem href={'href'} text={'text'} />
+const stories = storiesOf('LinkListItem', module)
+stories.addDecorator(withKnobs)
+stories.add('default', () => (
+  <LinkListItem
+    href={text('href', 'https://www.brianhan.co/')}
+    text={text('text', 'brianhan.co')}
+  />
 ))
