+++ src/utils/index.js
@@ -1,0 +1,12 @@
+import typography, { rem } from './typography'
+
+/**
+ * Reuse with defaultProps event handlers
+ *
+ * defaultProps = {
+ *   onClick: noop
+ * }
+ */
+const noop = () => {}
+
+export { noop, rem, typography }
